+++ libs/crispr-forms/src/lib/material/image-upload-field/image-upload-field.component.ts
@@ -9,7 +9,7 @@ import {
   Self,
   ChangeDetectorRef,
 } from '@angular/core';
-import { ControlValueAccessor, FormControl, NgControl, ReactiveFormsModule, ValidationErrors, ValidatorFn } from '@angular/forms';
+import { ControlValueAccessor, NgControl, ReactiveFormsModule } from '@angular/forms';
 import { MatButtonModule } from '@angular/material/button';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatIconModule } from '@angular/material/icon';
@@ -213,47 +213,3 @@ export class ImageUploadFieldComponent extends ImageUploadFieldMixin implements 
 }
 
 
-export function maxFileSizeValidator(maxFileSize: number): ValidatorFn {
-  return (control: FormControl): ValidationErrors | null => {
-    const controlValue = control.value;
-    if (!(controlValue instanceof File)) {
-      return null;
-    }
-    const actualFileSize = +convertBytesToMb(controlValue.size).toPrecision(3);
-    if (maxFileSize > actualFileSize) return null;
-    return {maxFileSize: { maxFileSize, actualFileSize }};
-  }
-}
-
-export function allowedFileExtValidator(allowedFileExtensions: string[], onlyAllowFiles = true): ValidatorFn {
-  return (control: FormControl): ValidationErrors | null => {
-    const controlValue = control.value;
-    const isFile = controlValue instanceof File;
-    if (!onlyAllowFiles && !isFile) return null;
-    const actualExtension = isFile ? controlValue.name.split('.').pop().toLocaleLowerCase() : 'Not a file';
-    const isAllowedExtension = allowedFileExtensions.includes(actualExtension);
-    if(isAllowedExtension) {
-      return null;
-    }
-    return {allowedExtension: { allowedFileExtensions, actualExtension }};
-  }
-}
-
-export function allowedFileType(matchExp: string | RegExp): ValidatorFn {
-  return (formControl) => {
-    const value = (formControl as FormControl).value;
-    if (!(value instanceof File)) return null
-    const type = value.type;
-    const matches = type.match(matchExp)
-    if (!type || !matches?.length) {
-      return {
-        allowedFileType: {
-          actualType: type,
-          allowedTypeExp: matchExp
-        }
-      }
-    } else {
-      return null
-    }
-  };
-}
