+++ libs/crispr-forms/src/lib/material/option/option.component.ts
@@ -1,5 +1,5 @@
 import { CommonModule } from '@angular/common';
-import { Component, Input, ChangeDetectionStrategy, NgModule } from '@angular/core';
+import { Component, Input, ChangeDetectionStrategy } from '@angular/core';
 import { SelectOption } from '../../models';
 import { InfoComponent } from '../info/info.component';
