+++ libs/crispr-forms/src/lib/models/checkbox-field.config.ts
@@ -1,5 +1,4 @@
-import { CrisprControlConfig, ControlType } from '../models';
-import { FieldDescriptors, MatFieldProperties } from './crispr-field.config';
+import { FieldDescriptors, MatFieldProperties, CrisprControlConfig, ControlType } from './crispr-field.config';
 
 export interface CheckboxFieldConfig extends
   CrisprControlConfig,
