+++ libs/crispr-forms/src/lib/models/divider.config.ts
@@ -1,4 +1,4 @@
-import { CrisprFieldConfig, ControlType } from '../models';
+import { CrisprFieldConfig, ControlType } from './crispr-field.config';
 // This is done different from the rest of the fields because
 // when we use the heading inside of a field we don't want to have to
 // pass a controlType
