+++ libs/crispr-forms/src/lib/models/heading.config.ts
@@ -1,5 +1,4 @@
-import { CrisprFieldConfig, ControlType } from '../models';
-import { FieldDescriptors } from './crispr-field.config';
+import { FieldDescriptors, CrisprFieldConfig, ControlType } from './crispr-field.config';
 // This is done different from the rest of the fields because
 // when we use the heading inside of a field we don't want to have to
 // pass a controlType
